% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci-class.R
\name{aci-class}
\alias{aci-class}
\alias{new_aci}
\alias{validate_aci}
\alias{aci}
\title{S3 class 'aci'}
\usage{
new_aci(x, id_cols = character(), covariate_cols = character())

validate_aci(x)

aci(x)
}
\arguments{
\item{x}{A data.frame or \link[tibble:tibble]{tibble} to be constructed into 'aci' class or validated.}

\item{id_cols}{A character string or vector of column names in \code{x} that are used to identify data (e.g. taxa, treatment, replicate, etc.). These will be converted to characters.}

\item{covariate_cols}{A character string or vector of column names in \code{x} that are used to identify covariates (numeric or categorical). Factors will be converted to characters.}
}
\description{
'aci' class objects are required for \code{\link{bayCi}}

\itemize{
\item \code{new_aci} constructs a data.frame into an 'aci' object
}

\itemize{
\item \code{validate_aci} validates that object has correct values
}

\itemize{
\item \code{aci} user-friendly way to create 'aci' objects
}
}
