% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayCi.R
\name{bayCi}
\alias{bayCi}
\title{Bayesian analysis of A-\eqn{Ci}{C_i} response curves using Stan}
\usage{
bayCi(formula, data)
}
\arguments{
\item{formula}{}

\item{data}{Data. Must inherit class \code{\link[=aci-class]{aci}}.}
}
\description{
Bayesian analysis of A-\eqn{Ci}{C_i} response curves using Stan
}
