% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baycir.R
\name{baycir}
\alias{baycir}
\title{Bayesian analysis of \deqn{A-C_i}{A-Ci} response curves using Stan}
\usage{
baycir(data)
}
\arguments{
\item{data}{Data. Must inherit class \code{\link[=acir-class]{acir}}.}
}
\description{
Bayesian analysis of \deqn{A-C_i}{A-Ci} response curves using Stan
}
