% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{braycir_model}
\alias{braycir_model}
\title{\strong{Stan} models for \strong{bayCi}}
\format{\strong{Stan} files}
\usage{
braycir_model
}
\description{
The \strong{Stan} code used for the different \strong{bayCi} models
}
\keyword{datasets}
