% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty-class.R
\name{empty-class}
\alias{empty-class}
\alias{new_empty}
\alias{validate_empty}
\alias{empty}
\title{S3 class 'empty'}
\usage{
new_empty(.x)

validate_empty(.x)

empty(.x)
}
\arguments{
\item{.x}{A \code{data.frame} or \code{\link[tibble:tibble]{tibble}} to be constructed into 'empty' class or validated.}
}
\description{
'empty' class objects are required for the empty argument in \code{\link{braycir}}

\itemize{
\item \code{new_empty} constructs a data.frame into an 'empty' object
}

\itemize{
\item \code{validate_empty} validates that object has correct values
}

\itemize{
\item \code{empty} user-friendly way to create 'empty' objects
}
}
