% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{get_c_constraints}
\alias{get_c_constraints}
\title{Constrain parameters on carboxylation-limited portion of the curve}
\usage{
get_c_constraints(data, gamma_star, Km, conf.level = 0.999)
}
\description{
Constrain parameters on carboxylation-limited portion of the curve
}
