% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{get_empty_constraints}
\alias{get_empty_constraints}
\title{Constrain parameters on empty curve correction}
\usage{
get_empty_constraints(empty, conf.level = 0.999)
}
\description{
Constrain parameters on empty curve correction
}
