% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{get_j_constraints}
\alias{get_j_constraints}
\title{Constrain parameters on RuBP-regeneration-limited portion of the curve}
\usage{
get_j_constraints(data, gamma_star, conf.level = 0.999)
}
\description{
Constrain parameters on RuBP-regeneration-limited portion of the curve
}
