% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiracir-class.R
\name{multiracir-class}
\alias{multiracir-class}
\alias{new_multiracir}
\alias{validate_multiracir}
\alias{multiracir}
\title{S3 class 'multiracir'}
\usage{
new_multiracir(.x, id_cols = character(), covariate_cols = character())

validate_multiracir(.x)

multiracir(.x)
}
\arguments{
\item{.x}{A \code{data.frame} or \code{\link[tibble:tibble]{tibble}} to be constructed into 'multiracir' class or validated.}

\item{id_cols}{A character string or vector of column names in \code{x} that are used to identify data (e.g. taxa, treatment, replicate, etc.). These will be converted to characters.}

\item{covariate_cols}{A character string or vector of column names in \code{x} that are used to identify covariates (numeric or categorical). Factors will be converted to characters.}
}
\description{
'multiracir' class objects are required for \code{\link{braycirs}}

\itemize{
\item \code{new_multiracir} constructs a data.frame into an 'multiracir' object
}

\itemize{
\item \code{validate_multiracir} validates that object has correct values
}

\itemize{
\item \code{multiracir} user-friendly way to create 'multiracir' objects
}
}
