% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_racirfit}
\alias{plot_racirfit}
\title{Plot A-Ci curve, data, and confidence bands from a \code{\link[racirfit-class]{racirfit}} object}
\usage{
plot_racirfit(racirfit, .width = 0.95, .point = stats::median,
  .interval = tidybayes::qi)
}
\arguments{
\item{racirfit}{An object of class \code{\link[=racirfit-class]{racirfit}} containing a fitted RACiR curve.}

\item{.width}{A probability to use that determine the width of the confidence intervals for plotting predicted A. Passed to \code{\link[tidybayes]{point_interval}}. Default is 0.95.}

\item{.point}{Point summary function for plotting predicted A from the posterior distribution. Function takes a vector and returns a single value, e.g. \code{\link[base]{mean}}, \code{\link[stats]{median}}, or \code{\link[tidybayes]{Mode}}. Default is \code{median}.}

\item{.interval}{Interval function, which takes a vector and a probability (\code{.width}) and returns a two-element vector representing the lower and upper bound of an interval; e.g. \code{\link[tidybayes]{qi}}, \code{\link[tidybayes]{hdi}}. Default is \code{qi}.}
}
\description{
Plot A-Ci curve, data, and confidence bands from a \code{\link[racirfit-class]{racirfit}} object
}
\seealso{
\code{\link[tidybayes]{point_interval}}, \code{\link[ggplot2]{ggplot}}
}
