% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_braycir}
\alias{predict_braycir}
\title{Predictions and confidence intervals from a \code{\link[racirfit-class]{racirfit}} object}
\usage{
predict_braycir(racirfit, .width = 0.95, .point = stats::median,
  .interval = tidybayes::qi)
}
\arguments{
\item{racirfit}{An object of class \code{\link[=racirfit-class]{racirfit}} containing a fitted RACiR curve.}

\item{.width}{A probability to use that determine the width of the confidence intervals for plotting predicted A. Passed to \code{\link[tidybayes]{point_interval}}. Default is 0.95.}

\item{.point}{Point summary function for plotting predicted A from the posterior distribution. Function takes a vector and returns a single value, e.g. \code{\link[base]{mean}}, \code{\link[stats]{median}}, or \code{\link[tidybayes]{Mode}}. Default is \code{median}.}

\item{.interval}{Interval function, which takes a vector and a probability (\code{.width}) and returns a two-element vector representing the lower and upper bound of an interval; e.g. \code{\link[tidybayes]{qi}}, \code{\link[tidybayes]{hdi}}. Default is \code{qi}.}
}
\value{
A \code{\link[tibble]{tibble}} with columns:

\itemize{
\item{\strong{Ci}: Corrected Ci value (\eqn{\mu \text{mol}~\text{mol}^{-1}}{umol / mol})}
\item{\strong{A}: Corrected A value (\eqn{\mu \text{mol}~\text{m}^{-2}~\text{s}^{-1}}{umol / m^2 / s})}
\item{\strong{lower} Lower confidence interval on \strong{A}}
\item{\strong{upper} Upper confidence interval on \strong{A}}
}

To make plotting easier, values are not assigned \code{\link[units]{units}}
}
\description{
Predictions and confidence intervals from a \code{\link[racirfit-class]{racirfit}} object
}
