% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racir-class.R
\name{racir-class}
\alias{racir-class}
\alias{new_racir}
\alias{validate_racir}
\alias{racir}
\title{S3 class 'racir'}
\usage{
new_racir(.x)

validate_racir(.x)

racir(.x)
}
\arguments{
\item{.x}{A \code{data.frame} or \code{\link[tibble:tibble]{tibble}} to be constructed into 'racir' class or validated.}
}
\description{
'racir' class objects are required for the data argument in \code{\link{braycir}}

\itemize{
\item \code{new_racir} constructs a \code{data.frame} into a 'racir' object
}

\itemize{
\item \code{validate_racir} validates that object has correct values
}

\itemize{
\item \code{racir} user-friendly way to create 'racir' objects
}
}
