% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racirfit-class.R
\name{racirfit-class}
\alias{racirfit-class}
\alias{new_racirfit}
\alias{validate_racirfit}
\alias{racirfit}
\title{S3 class 'racirfit'}
\usage{
new_racirfit(.x)

validate_racirfit(.x)

racirfit(.x)
}
\arguments{
\item{.x}{A \code{list} to be constructed into 'racirfit' class or validated.}
}
\value{
\describe{
\item{data}{A \code{\link[=racir-class]{racir}} object}
\item{empty}{A \code{\link[=empty-class]{empty}} object with additional information on which data points were used for correction}
\item{model}{The model code in \strong{Stan} language}
\item{R2}{The Bayesian R-squared value}
\item{fit}{An object of class \code{\link[rstan]{stanfit}} among others containing the posterior samples}
\item{version}{The versions of \strong{bayCi} and \strong{rstan} with which the model was fitted}
}
}
\description{
The components of a \code{racirfit} object and the various methods are described below. When methods have their own more detailed documentation pages links are provided.

\itemize{
\item \code{new_racirfit} constructs a list into an 'racir' object
}

\itemize{
\item \code{validate_racirfit} validates that object has correct values
}

\itemize{
\item \code{racirfit} user-friendly way to create 'racirfit' objects
}
}
\seealso{
\link[rstan]{stan}
}
