% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_aci}
\alias{read_aci}
\title{Read LICOR files into R}
\usage{
read_aci(path)
}
\arguments{
\item{path}{Character strings or vector of character stings indicating LICOR files to be read (.csv, .xls, or .xlsx). Single files can be imported, or multiple files can be imported and concatenated by row. If a directory name is given, all files in the directory will be read and concatenated by row.}
}
\description{
Read LICOR files into R
}
\examples{

library(bayCi)

}
