% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_aci_file}
\alias{read_aci_file}
\title{Read single file with one or more A-Ci curves}
\usage{
read_aci_file(file, ...)
}
\arguments{
\item{file}{Character string with path to file

This function is not meant to be called directly. \code{read_aci} will have already checked that file exists and has valid extension.

#' @noRd}
}
\description{
Read single file with one or more A-Ci curves
}
