% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate.R
\name{recalculate}
\alias{recalculate}
\alias{recalc_licor68}
\title{Recalculate LI-6800 values}
\usage{
recalc_licor68(.x)
}
\arguments{
\item{.x}{A \code{data.frame} containing output from the LI-6800.}
}
\description{
Recalculate LI-6800 values
}
\examples{

library(bayCi)

}
