% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-model.R
\name{write_braycir_model}
\alias{write_braycir_model}
\title{Write \code{braycir} model}
\usage{
write_braycir_model(data, empty, gamma_star, Km)
}
\arguments{
\item{data}{An object of class \code{\link[=racir-class]{racir}} containing data for all model variables.}

\item{empty}{An object of class \code{\link[=empty-class]{empty}} containing data from empty chamber to correct response data.}

\item{gamma_star}{this is probably going to change}

\item{Km}{this is probably going to change}
}
\description{
Write \code{braycir} model
}
