% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acir-class.R
\name{acir-class}
\alias{acir-class}
\alias{new_acir}
\alias{validate_acir}
\alias{acir}
\title{S3 class 'acir'}
\usage{
new_acir(x)

validate_acir(x)

acir(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{\link[tibble:tibble]{tibble}} to be constructed into 'acir' class or validated.}
}
\description{
'acir' class objects are required for \code{\link{baycir}}

\itemize{
\item \code{new_acir} constructs a data.frame into an 'acir' object
}

\itemize{
\item \code{validate_acir} validates that object has correct values
}

\itemize{
\item \code{acir} user-friendly way to create 'acir' objects
}
}
