% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayCi.R
\docType{package}
\name{bayCi}
\alias{bayCi}
\alias{bayCi-package}
\title{\code{bayCi} package}
\description{
Bayesian Photosynthetic Response Curves Using Stan
}
\details{
See the README on
\href{https://github.com/cdmuir/bayCi}{GitHub}
}
